requirements.txt
fastapi==0.115.0
sqlalchemy==2.0.35
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.4
wtforms==3.1.2
uvicorn==0.30.6